refactor(time): consolidate time utilities into atuin-common
Move duration formatting, the datetime format descriptors, Timezone and the
local offset helper into atuin_common::time, split into duration,
offset_date_time and timezone modules.

Adds try_unix_nanos_{i64,u64}, saturating_from_nanos_i64 and
saturating_duration_since, replacing hand-rolled equivalents across 8 call
sites. Timezone is re-exported from settings, so callers are unchanged.

Also fixes:
- the daemon expect()ed on the end_history duration conversion, and could
  persist a negative duration when the clock moved backwards
- history_list re-parsed its datetime format descriptor on every row render
