+++ crates/atuin-daemon/src/components/history.rs
@@ -190,11 +190,15 @@ impl HistorySvc for HistoryGrpcService {
         if let Some((_, mut history)) = self.inner.running.remove(&id) {
             history.exit = req.exit;
             history.duration = match req.duration {
+                // saturating: a clock that moved backwards must not persist a negative
+                // duration, and must not take the daemon down either
                 0 => i64::try_from(
-                    (OffsetDateTime::now_utc() - history.timestamp).whole_nanoseconds(),
+                    OffsetDateTime::now_utc()
+                        .saturating_duration_since(history.timestamp)
+                        .as_nanos(),
                 )
-                .expect("failed to convert calculated duration to i64"),
-                value => i64::try_from(value).expect("failed to get i64 duration"),
+                .unwrap_or(i64::MAX),
+                value => i64::try_from(value).unwrap_or(i64::MAX),
             };
 
             // Get the handle and store to save the history
